added Edit modes to view and model and keyboard shortcuts to select them.
Also switched convention that a mode is enabled (e.g., show potential helix) by clicking a key once and disabled by clicking it again. Before the convention was a mode is enabled while a non-control key is pressed, but it turns out touchpads become disabled while keys are pressed, so it was impossible to press a key and then click somewhere. (Still possible with control keys such as Ctrl and Shift).