Fix/Add unit tests for dna end bug fixes
After fixing issue #77 with commit

https://github.com/UC-Davis-molecular-computing/scadnano/commit/45f09682b8f2a71ed0941e573f86719869cda890

previous unit tests fail for having null fields `lowest_offset`
and `highest_offset`.

After fixing #72 and #83, DNAEndsMove no longer updates the
UndoRedo stack, so the DNAEndsMove needs to be fixed.

The unit test for #77 did not change the model so there was some
ambiguity over whether the UndoRedo stack
and change_since_last_saved field should be updated.
This commit creates a new unit test where the model is updated
to remove this ambiguity.

Lastly, unit tests are added for #83