Fixes unit tests for strand move
https://github.com/UC-Davis-molecular-computing/scadnano/commit/b44ca3d412ea734bc55aca6ecf10ad04e7412382

The above commit change the API of strand move actions,
so a rewrite of unit tests is needed.

A unit test is included for the bug noted in #114

Unit test for #50 have been removed since the invalid action
is never constructed at all.

Basic tests for adding insertion and deletions as well
as grid changes are also added.