Adds strands recoloring for comparison
A previous commit removed the BuiltValue annotation
for ignoring the strand color field for hashing and
comparisons, causing unit tests to fail. This commit
adds a utility function to recolor strands.

A previous commit also changed the StrandsMoveStart
api by now requiring a strands field. This commit
adds this field to the unit tests.