Replaces png with svg when exporting main
Fixes #216

Previously if the dna sequence png is being used and the user exports main,
then the output svg file would also use the png image, which is not the
expected output. The dna sequence svg should be used instead of the png.

Implementation Summary

The UI state now includes a field named disable_png_cache_until_action_completes,
which holds any action (in this case, exporting main), that needs png to be
disabled before being dispatched. This state field tells the view to dispatch
the held action, and this is accomplished by overriding React's
componentDidUpdate function for DesignMainDNASequencesComponent. The view
tells the state to re-enable png caching by dispatching a
setDisablePngCacheUntilActionCompletes action. The export svg middleware also
dispatches this action to tell the state to disable png caching and dispatch
the export main action later.

Details

* lib/src/actions/actions.dart
  * Adds setDisablePngCacheUntilActionCompletes (ui_state setter)
* lib/src/middleware/export_svg.dart
  * Adds logic for dispatching setDisablePngCacheUntilActionCompletes
* lib/src/reducers/app_ui_state_reducer.dart
  * Adds setDisablePngCacheUntilActionCompletes reducer
* lib/src/serializers.dart
  * Adds setDisablePngCacheUntilActionCompletes serializer
* lib/src/state/app_ui_state.dart
  * Adds `disable_png_cache_until_action_completes` field to `AppUIState`
* lib/src/util.dart
  * Adds use_png utility function
* lib/src/view/design_main.dart
  * Adds disable_png_cache_until_action_completes prop
* lib/src/view/design_main_dna_sequences.dart
  * Adds componentDidUpdate function to handle
  `disable_png_cache_until_action_completes` prop.
  * Slight code cleanup using new utility function from util.dart
* test/reducer_test.dart
  * Adds setDisablePngCacheUntilActionCompletes test